refactor entity_ids, and consolidate classes
