+++ homeassistant/components/incomfort/__init__.py
@@ -1,5 +1,6 @@
 """Support for an Intergas boiler via an InComfort/Intouch Lan2RF gateway."""
 import logging
+from typing import Optional
 
 from aiohttp import ClientResponseError
 import voluptuous as vol
@@ -8,7 +9,9 @@ from incomfortclient import Gateway as InComfortGateway
 from homeassistant.const import CONF_HOST, CONF_PASSWORD, CONF_USERNAME
 from homeassistant.helpers import config_validation as cv
 from homeassistant.helpers.aiohttp_client import async_get_clientsession
+from homeassistant.core import callback
 from homeassistant.helpers.discovery import async_load_platform
+from homeassistant.helpers.dispatcher import async_dispatcher_connect
 
 _LOGGER = logging.getLogger(__name__)
 
@@ -53,3 +56,44 @@ async def async_setup(hass, hass_config):
         )
 
     return True
+
+
+class IncomfortEntity:
+    """Base class for all InComfort entities."""
+
+    def __init__(self) -> None:
+        """Initialize the class."""
+        self._unique_id = self._name = None
+
+    @property
+    def unique_id(self) -> Optional[str]:
+        """Return a unique ID."""
+        return self._unique_id
+
+    @property
+    def name(self) -> Optional[str]:
+        """Return the name of the sensor."""
+        return self._name
+
+
+class IncomfortChild(IncomfortEntity):
+    """Base class for all InComfort entities (excluding the boiler)."""
+
+    def __init__(self) -> None:
+        """Initialize the class."""
+        super().__init__()
+
+    async def async_added_to_hass(self) -> None:
+        """Set up a listener when this entity is added to HA."""
+        # pylint: disable=no-member
+        async_dispatcher_connect(self.hass, DOMAIN, self._refresh)
+
+    @callback
+    def _refresh(self) -> None:
+        # pylint: disable=no-member
+        self.async_schedule_update_ha_state(force_refresh=True)
+
+    @property
+    def should_poll(self) -> bool:
+        """Return False as this device should never be polled."""
+        return False
