+++ homeassistant/components/incomfort/binary_sensor.py
@@ -1,11 +1,9 @@
 """Support for an Intergas heater via an InComfort/InTouch Lan2RF gateway."""
 from typing import Any, Dict, Optional
 
-from homeassistant.components.binary_sensor import BinarySensorDevice
-from homeassistant.core import callback
-from homeassistant.helpers.dispatcher import async_dispatcher_connect
+from homeassistant.components.binary_sensor import BinarySensorDevice, ENTITY_ID_FORMAT
 
-from . import DOMAIN
+from . import DOMAIN, IncomfortChild
 
 
 async def async_setup_platform(hass, config, async_add_entities, discovery_info=None):
@@ -18,33 +16,19 @@ async def async_setup_platform(hass, config, async_add_entities, discovery_info=
     )
 
 
-class IncomfortFailed(BinarySensorDevice):
+class IncomfortFailed(IncomfortChild, BinarySensorDevice):
     """Representation of an InComfort Failed sensor."""
 
     def __init__(self, client, heater) -> None:
         """Initialize the binary sensor."""
+        super().__init__()
+
         self._unique_id = f"{heater.serial_no}_failed"
+        self.entity_id = ENTITY_ID_FORMAT.format(f"{DOMAIN}_failed")
+        self._name = "InComfort Fault state"
 
         self._client = client
         self._heater = heater
-
-    async def async_added_to_hass(self) -> None:
-        """Set up a listener when this entity is added to HA."""
-        async_dispatcher_connect(self.hass, DOMAIN, self._refresh)
-
-    @callback
-    def _refresh(self) -> None:
-        self.async_schedule_update_ha_state(force_refresh=True)
-
-    @property
-    def unique_id(self) -> Optional[str]:
-        """Return a unique ID."""
-        return self._unique_id
-
-    @property
-    def name(self) -> Optional[str]:
-        """Return the name of the sensor."""
-        return "Fault state"
 
     @property
     def is_on(self) -> bool:
@@ -55,8 +39,3 @@ class IncomfortFailed(BinarySensorDevice):
     def device_state_attributes(self) -> Optional[Dict[str, Any]]:
         """Return the device state attributes."""
         return {"fault_code": self._heater.status["fault_code"]}
-
-    @property
-    def should_poll(self) -> bool:
-        """Return False as this device should never be polled."""
-        return False
