+++ homeassistant/components/incomfort/climate.py
@@ -1,16 +1,14 @@
 """Support for an Intergas boiler via an InComfort/InTouch Lan2RF gateway."""
 from typing import Any, Dict, List, Optional
 
-from homeassistant.components.climate import ClimateDevice
+from homeassistant.components.climate import ClimateDevice, ENTITY_ID_FORMAT
 from homeassistant.components.climate.const import (
     HVAC_MODE_HEAT,
     SUPPORT_TARGET_TEMPERATURE,
 )
 from homeassistant.const import ATTR_TEMPERATURE, TEMP_CELSIUS
-from homeassistant.core import callback
-from homeassistant.helpers.dispatcher import async_dispatcher_connect
 
-from . import DOMAIN
+from . import DOMAIN, IncomfortChild
 
 
 async def async_setup_platform(hass, config, async_add_entities, discovery_info=None):
@@ -24,39 +22,19 @@ async def async_setup_platform(hass, config, async_add_entities, discovery_info=
     async_add_entities([InComfortClimate(client, heater, r) for r in heater.rooms])
 
 
-class InComfortClimate(ClimateDevice):
+class InComfortClimate(IncomfortChild, ClimateDevice):
     """Representation of an InComfort/InTouch climate device."""
 
     def __init__(self, client, heater, room) -> None:
         """Initialize the climate device."""
+        super().__init__()
+
         self._unique_id = f"{heater.serial_no}_{room.room_no}"
+        self.entity_id = ENTITY_ID_FORMAT.format(f"{DOMAIN}_{room.room_no}")
+        self._name = f"Thermostat {room.room_no}"
 
         self._client = client
         self._room = room
-        self._name = f"Room {room.room_no}"
-
-    async def async_added_to_hass(self) -> None:
-        """Set up a listener when this entity is added to HA."""
-        async_dispatcher_connect(self.hass, DOMAIN, self._refresh)
-
-    @callback
-    def _refresh(self) -> None:
-        self.async_schedule_update_ha_state(force_refresh=True)
-
-    @property
-    def should_poll(self) -> bool:
-        """Return False as this device should never be polled."""
-        return False
-
-    @property
-    def unique_id(self) -> Optional[str]:
-        """Return a unique ID."""
-        return self._unique_id
-
-    @property
-    def name(self) -> str:
-        """Return the name of the climate device."""
-        return self._name
 
     @property
     def device_state_attributes(self) -> Dict[str, Any]:
