+++ homeassistant/components/incomfort/sensor.py
@@ -1,18 +1,17 @@
 """Support for an Intergas heater via an InComfort/InTouch Lan2RF gateway."""
 from typing import Any, Dict, Optional
 
+from homeassistant.components.sensor import ENTITY_ID_FORMAT
 from homeassistant.const import (
     PRESSURE_BAR,
     TEMP_CELSIUS,
     DEVICE_CLASS_PRESSURE,
     DEVICE_CLASS_TEMPERATURE,
 )
-from homeassistant.core import callback
-from homeassistant.helpers.dispatcher import async_dispatcher_connect
 from homeassistant.helpers.entity import Entity
 from homeassistant.util import slugify
 
-from . import DOMAIN
+from . import DOMAIN, IncomfortChild
 
 INCOMFORT_HEATER_TEMP = "CV Temp"
 INCOMFORT_PRESSURE = "CV Pressure"
@@ -42,37 +41,22 @@ async def async_setup_platform(hass, config, async_add_entities, discovery_info=
     )
 
 
-class IncomfortSensor(Entity):
+class IncomfortSensor(IncomfortChild, Entity):
     """Representation of an InComfort/InTouch sensor device."""
 
     def __init__(self, client, heater, name) -> None:
         """Initialize the sensor."""
+        super().__init__()
+
         self._client = client
         self._heater = heater
 
         self._unique_id = f"{heater.serial_no}_{slugify(name)}"
-
+        self.entity_id = ENTITY_ID_FORMAT.format(f"{DOMAIN}_{slugify(name)}")
         self._name = name
+
         self._device_class = None
         self._unit_of_measurement = None
-
-    async def async_added_to_hass(self) -> None:
-        """Set up a listener when this entity is added to HA."""
-        async_dispatcher_connect(self.hass, DOMAIN, self._refresh)
-
-    @callback
-    def _refresh(self) -> None:
-        self.async_schedule_update_ha_state(force_refresh=True)
-
-    @property
-    def unique_id(self) -> Optional[str]:
-        """Return a unique ID."""
-        return self._unique_id
-
-    @property
-    def name(self) -> Optional[str]:
-        """Return the name of the sensor."""
-        return self._name
 
     @property
     def state(self) -> Optional[str]:
@@ -88,11 +72,6 @@ class IncomfortSensor(Entity):
     def unit_of_measurement(self) -> Optional[str]:
         """Return the unit of measurement of the sensor."""
         return self._unit_of_measurement
-
-    @property
-    def should_poll(self) -> bool:
-        """Return False as this device should never be polled."""
-        return False
 
 
 class IncomfortPressure(IncomfortSensor):
